Fix GPU logical signature tracker lookup
The GPU kernel was not checking tracker_type when looking up logical
signature trackers. This caused the kernel to fail to find the correct
tracker for evaluating multi-part logical signatures.

Logical signature trackers have tracker_type=1, while multi-part pattern
trackers have tracker_type=0. The lookup now correctly filters by both
sig_id AND tracker_type to ensure the right tracker is found.

This fixes the issue where logical signatures would fail to evaluate
properly on the GPU, preventing detection of viruses that match complex
logical expression patterns.

+++ libclamav/matcher-ac.h
@@ -108,7 +108,7 @@ struct cli_ac_patt {
     uint32_t boundary;
     uint8_t depth;
     uint8_t sigopts;
-    uint32_t gpu_id;
+    uint16_t gpu_id;
 };
 
 struct cli_ac_list {
